requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pytz==2023.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
wheel==0.44.0
